Remove attempt at `->toArray()` on passed object
Tried to be too clever in the past by calling `->toArray()` on objects passed in the view data if the method existed.

it never really worked well (not being a deep traversal) and added a tiny bit of unnecessary magic, which was confusing as hell if stepping back into a project that used this package.